Add a storm REST API agent
This commit adds two things - a cron job-able script to scrape the Storm
REST API and cache the resulting data as a JSON file, and a check that
takes the same JSON file & generates metrics off it, with an additional
set of tags provided per-task from another JSON file.

The two scripts share a config file and as much code as makes sense to
share, which should help us keep them in sync. (:

In my tests, it takes about 20-30 seconds to generate the cache file,
and like 3 seconds to generate 10k metrics from the cached JSON.