Don't scale microsecond storm latency metrics
Storm reports them in the least-useful format in the world:
microseconds, printed as float (so you can get
pooooosssssibly-nanoseconds with meh precision?). Initially assuming
they were reported in seconds, I scaled them up by 1000 to get
milliseconds; that's not right.

We should just not scale them and report them as the unit they really
are, instead.