Emit monotonic counts for tuple counts
We emitted them as rates, but sadly we're getting them as counters
across the lifetime of the topology. This should make DD's UI treat the
thing as a count and not a gauge (I was hoping we could fix this by
setting them as counters in the UI, but there are Bugs).